--- AppResources_rc.py
@@ -1,93 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Resource object code
-#
-# Created by: The Resource Compiler for PyQt4 (Qt v4.8.7)
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt4 import QtCore
-
-qt_resource_data = b"\
-\x00\x00\x03\x0e\
-\x89\
-\x50\x4e\x47\x0d\x0a\x1a\x0a\x00\x00\x00\x0d\x49\x48\x44\x52\x00\
-\x00\x00\x5e\x00\x00\x00\x80\x08\x06\x00\x00\x00\x82\xe6\xde\x31\
-\x00\x00\x00\x04\x73\x42\x49\x54\x08\x08\x08\x08\x7c\x08\x64\x88\
-\x00\x00\x00\x09\x70\x48\x59\x73\x00\x00\x0b\x13\x00\x00\x0b\x13\
-\x01\x00\x9a\x9c\x18\x00\x00\x02\xb0\x49\x44\x41\x54\x78\x9c\xed\
-\xdd\x49\x4e\x1b\x41\x00\x85\xe1\xe7\x88\x1b\x64\x6f\x33\x1a\x83\
-\xc1\x60\xe0\x44\x59\xe5\x44\x51\x2e\x17\x42\xe6\x39\x51\x0e\x80\
-\xc8\xa2\xba\xa3\x4e\x64\x3c\x74\x77\xd5\xeb\x76\xfd\xbf\xf4\x36\
-\xc8\xd0\x55\x9f\x58\x19\xc9\x48\x44\x44\xb4\xa4\x9b\x62\x94\xb0\
-\x1b\x49\xbf\x8a\x5d\x9b\xcf\x92\x4d\xd7\x0a\xe0\x0f\xc5\x7e\x4a\
-\xba\xb2\x9e\x28\x83\xae\x14\xa0\x1f\xfe\x1b\xf8\x11\x7b\x0c\x1d\
-\xfc\x88\xad\x42\xaf\xe2\xcf\x4d\x67\xdc\xba\xe6\x5a\x0f\xbd\xdc\
-\x0f\x81\xdf\xb8\xb9\x02\xe4\xba\xe8\x55\xfc\x4b\xc3\x79\xb7\xa2\
-\x4b\xd5\x43\x2f\xf7\x5d\xd2\x45\xf2\x53\xf7\xbc\x0b\x05\xb8\xba\
-\xe8\xe0\xd7\xa8\x2d\xf4\x2a\xfe\x2c\xe9\x0d\x7a\xd8\x4c\xed\xa2\
-\x97\xfb\x26\xf0\x1f\x2d\x16\x3a\xf8\x4b\x9a\x29\xc0\xc4\x42\xaf\
-\xe2\x9f\x27\xba\x53\xe7\x3b\x57\x1a\xf4\x72\x5f\x25\x9d\x25\xb9\
-\x59\x87\x3b\x53\x80\x48\x85\x0e\xbe\x7c\xe8\x55\xfc\x69\xf4\x5b\
-\x76\xac\xa9\xbc\xe8\xe5\xbe\x28\x23\xfc\xae\xa0\x67\x85\x3f\x55\
-\xb8\xa8\x1b\x7b\x11\xfe\x69\xc4\x7b\x5b\x3b\x55\x37\xd1\xcb\x7d\
-\x96\x74\x12\xed\xf6\xa6\x4e\x14\x2e\xe6\xc6\xcd\x0a\xbf\x2f\xe8\
-\xe5\x3e\x49\x9a\x44\x91\x48\xd8\x44\xe1\x22\x6e\xcc\xac\xf0\xfb\
-\x8a\x5e\xc5\x3f\x6e\x5d\x25\x72\x7d\x47\xef\x25\xfe\xb1\xb6\x03\
-\xbd\xdc\x47\xf5\x00\xff\x58\xe1\xa0\x6e\xac\x18\xf8\xe3\x16\x9d\
-\x5a\x6d\xac\xed\x44\x2f\xf7\x41\x1d\xc4\x1f\x2b\x1c\xcc\x8d\x93\
-\x02\xff\xa8\x25\xb3\xc6\x1d\x29\x0f\xf4\x72\xef\xd5\x01\xfc\xdc\
-\xd0\xab\xf8\x87\x2d\xf8\xd5\xea\xb0\x38\x80\x1b\x21\x2b\xfc\xdc\
-\xd1\xcb\xbd\x93\x74\xd0\xd0\x72\xed\x0e\x8a\x07\xba\x2f\xdd\x95\
-\x25\xc1\x07\x7d\xf1\xde\x4a\xda\x6f\xe0\xba\xb4\xfd\xe2\x01\xee\
-\x4b\x76\x75\x51\xf0\x41\x5f\x6f\x6f\xd4\x22\x3e\xe8\x9b\xe3\xef\
-\xd5\x92\xae\xb4\x57\xfc\x20\xf7\x65\xfa\xb6\x46\xf8\xa0\x37\xdb\
-\x9d\xa4\xdd\x0d\xcd\xb5\x5b\x7c\xa3\xfb\xf0\x7d\xdf\x9d\x36\xc0\
-\x07\xbd\x7d\xfc\xd1\x2a\xf4\x11\xe8\x51\xf6\x7a\x19\xfe\xa8\x78\
-\x81\xfb\x90\xdb\xba\xdb\x45\xf8\xa0\xa7\xc3\x1f\x96\xe8\xc3\xe2\
-\x0b\xee\x43\xe5\xb2\x5b\x49\xc3\x81\xa4\x67\x8a\xf8\x3e\x43\xa5\
-\xa7\x92\x9e\x27\x78\x4e\x93\x5e\x28\x7c\x40\x45\xec\x5e\x25\x78\
-\xc6\xdf\x26\xf2\xff\xb6\xad\x5a\xb2\xf7\xd7\x9f\xa4\x7a\x10\xfd\
-\x1b\xf0\xa6\x80\x37\x05\xbc\x29\xe0\x4d\x01\x6f\x0a\x78\x53\xc0\
-\x9b\x02\xde\x14\xf0\xa6\x80\x37\x05\xbc\x29\xe0\x4d\x01\x6f\x0a\
-\x78\x53\xc0\x9b\x02\xde\x14\xf0\xa6\x80\x37\x05\xbc\x29\xe0\x4d\
-\x01\x6f\x0a\x78\x53\xc0\x9b\x02\xde\x14\xf0\xa6\x80\x37\x05\xbc\
-\x29\xe0\x4d\x01\x6f\x0a\x78\x53\xc0\x9b\x02\xde\x14\xf0\xa6\x80\
-\x37\x05\xbc\x29\xe0\x4d\x01\x6f\x6a\x27\xe1\xb3\xee\x25\xfd\x4e\
-\xf8\xbc\x3a\xdd\xbb\x0f\x40\x44\x44\x44\x44\x44\x44\x44\x44\x94\
-\x4d\x03\x49\x2f\xdd\x87\xc8\xb1\x81\xc2\xa7\xcf\x51\xe2\xf8\x43\
-\x88\x29\xe0\x4d\x01\x6f\x0a\x78\x53\xc0\x9b\x02\xde\x14\xf0\xa6\
-\x80\x37\x05\xbc\x29\xe0\x4d\x01\x6f\x6a\x47\xe1\x5f\x3f\x13\x11\
-\x11\x11\x11\x11\x11\x11\x55\xfa\x03\x23\x67\xa1\x95\x53\x1d\x9f\
-\x0f\x00\x00\x00\x00\x49\x45\x4e\x44\xae\x42\x60\x82\
-"
-
-qt_resource_name = b"\
-\x00\x05\
-\x00\x6f\xa6\x53\
-\x00\x69\
-\x00\x63\x00\x6f\x00\x6e\x00\x73\
-\x00\x03\
-\x00\x00\x70\x37\
-\x00\x69\
-\x00\x6d\x00\x67\
-\x00\x0f\
-\x01\x0e\x6d\x07\
-\x00\x75\
-\x00\x70\x00\x6c\x00\x6f\x00\x61\x00\x64\x00\x5f\x00\x69\x00\x63\x00\x6f\x00\x6e\x00\x2e\x00\x70\x00\x6e\x00\x67\
-"
-
-qt_resource_struct = b"\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
-\x00\x00\x00\x10\x00\x02\x00\x00\x00\x01\x00\x00\x00\x03\
-\x00\x00\x00\x1c\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-"
-
-def qInitResources():
-    QtCore.qRegisterResourceData(0x01, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-def qCleanupResources():
-    QtCore.qUnregisterResourceData(0x01, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-qInitResources()
