+++ res/README.md
@@ -1,0 +1,8 @@
+# VID-GUI
+
+This is a readme file to inspect the recent updates released by Shashank Sharma
+
+1. Added Sync to all videos by compromising 0.4 second delay
+2. Added Terminal Header
+3. Removed Segmentation Fault Error
+4. Tested with 360p 10MB video
